chapter9/1_src/requirements.txt
numpy
Pillow
tensorflow